Rearrange molecules folder hierarchy
