+++ src/components/molecules/buttons/overlayedIcon.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 
 import { makeStyles } from '@material-ui/core';
 
-import IconButton from '../atoms/buttons/icon.jsx';
-import Footer from './footer.jsx';
+import IconButton from '../../atoms/buttons/icon.jsx';
+import Footer from '../layout/footer.jsx';
 
 export default function OverlayedIconButton({ color, children, handleClick }) {
   const useStyles = makeStyles(theme => ({
