+++ src/components/molecules/fields/labeledCheckbox.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 
 import { makeStyles } from '@material-ui/core';
 
-import Typography from '../atoms/typography.jsx';
-import Checkbox from '../atoms/fields/checkbox.jsx';
+import Typography from '../../atoms/typography.jsx';
+import Checkbox from '../../atoms/fields/checkbox.jsx';
 
 export default function LabeledCheckbox({ id, value, handleChange, children }) {
   const useStyles = makeStyles({
