+++ src/components/molecules/layout/footer.stories.jsx
@@ -5,7 +5,7 @@ import { storiesOf } from '@storybook/react';
 import theme from '../../../styles/global-styles';
 import themeDecorator from '../../../../.storybook/decorators/themeDecorator';
 
-import Footer from '../footer';
+import Footer from './footer';
 
 storiesOf('Molecules/Layout/Footer', module)
   .addDecorator(themeDecorator(theme))
