+++ src/components/molecules/recipe/list/toolbar.jsx
@@ -1,23 +1,31 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Grid, Divider } from '@material-ui/core';
-import { NotificationsOutlined as NotificationsIcon, Check as CheckIcon } from '@material-ui/icons';
+import { makeStyles, Grid, Divider } from '@material-ui/core';
 
-import IconButton from '../atoms/buttons/icon.jsx';
-import TextField from '../atoms/fields/text.jsx';
-import About from './about.jsx';
-import DrawerPanel from '../organisms/drawerPanel.jsx';
+import TextField from '../../../atoms/fields/text.jsx';
+import About from '../../about.jsx';
+import DrawerPanel from '../../../organisms/drawerPanel.jsx';
 
-export default function ItemListToolbar({
+export default function RecipeListToolbar({
   authorData,
   projectData,
   logo,
   className,
   searchValue,
   handleChange,
-  handleClick,
+  literals,
 }) {
+  const useStyles = makeStyles({
+    description: {
+      width: '300px',
+    },
+    alignText: {
+      'text-align': 'center',
+    }
+  });
+  const classes = useStyles();
+
   return (
     <div className={className}>
       <Grid container alignItems="center">
@@ -31,44 +39,43 @@ export default function ItemListToolbar({
               projectUrl={projectData.url}
             />
             <Divider />
+            <div className={`${classes.description} ${classes.alignText}`}>
+              <p>{projectData.description}</p>
+              <p>
+                <strong>{literals.participants}</strong>
+              </p>
+              {projectData.participants.map(
+                (name) => <p key={`${name}`}>{name}</p>
+              )}
+            </div>
           </DrawerPanel>
         </Grid>
-        <Grid item xs={6} container justify="center">
+        <Grid item xs={8} container justify="center">
           <form noValidate autoComplete="off">
             <TextField id="search" value={searchValue} handleChange={handleChange} />
           </form>
-        </Grid>
-        <Grid item xs={2} container justify="flex-end">
-          <IconButton handleClick={handleClick}>
-            <NotificationsIcon />
-          </IconButton>
-        </Grid>
-        <Grid item xs={2} container justify="flex-start">
-          <IconButton handleClick={handleClick}>
-            <CheckIcon />
-          </IconButton>
         </Grid>
       </Grid>
     </div>
   );
 }
 
-ItemListToolbar.defaultProps = {
+RecipeListToolbar.defaultProps = {
   authorData: {},
   projectData: {},
   className: '',
   searchValue: null,
   handleChange: () => {},
-  handleClick: () => {},
 };
 
-ItemListToolbar.propTypes = {
+RecipeListToolbar.propTypes = {
   authorData: PropTypes.shape({
     name: PropTypes.string,
     email: PropTypes.string,
     url: PropTypes.string,
   }),
   projectData: PropTypes.shape({
+    description: PropTypes.string,
     logo: PropTypes.string,
     url: PropTypes.string,
   }),
@@ -76,5 +83,4 @@ ItemListToolbar.propTypes = {
   className: PropTypes.string,
   searchValue: PropTypes.string,
   handleChange: PropTypes.func,
-  handleClick: PropTypes.func,
 };
