+++ src/components/organisms/itemDetailFooterBar.jsx
@@ -5,7 +5,7 @@ import { makeStyles, Grid } from '@material-ui/core';
 
 import Typography from '../atoms/typography.jsx';
 import TextButton from '../atoms/buttons/text.jsx';
-import Footer from '../molecules/footer.jsx';
+import Footer from '../molecules/layout/footer.jsx';
 
 export default function ItemDetailFooterBar({ literals, data, handleClick }) {
   const useStyles = makeStyles(theme => ({
